apps/phoenix-backend-unified/requirements.txt
# 🔐 Phoenix Backend Unified - Enterprise Auth Hub
# Luna Session Zero Complete Authentication System

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==23.0.0

# Database & ORM
supabase==2.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Configuration & Environment
python-dotenv==1.0.0

# HTTP Client
httpx==0.27.0

# Utilities
aiofiles==23.2.1

# Validation
pydantic[email]==2.5.2

# Date/Time
python-dateutil==2.8.2